requirements.txt
pytest==7.2.1
playwright==1.31.0
allure-pytest==2.13.2